requirements.txt
# Core dependencies
numpy>=1.24.0
openai>=1.0.0  # Optional - for OpenAI embeddings
requests>=2.32.0
pyyaml>=6.0.0

# Database
# SQLite is included in Python standard library

# Logging and utilities
# All other utilities are standard library

